Multiple features finally added!
it will now pull patches links, it will not complain if you don't have
that in the pkg.json, another huge thing is that the build_dir is now
available! Furthermore, quality of life changes like the directories
actually working, so like cd will actually cd, and everything is bing
chilling. This is finally in the PREVIEW stages, I will release a
tarball soon to download, and change the Makefile, since it's finally
somewhat usable.